backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot
google-generativeai
python-multipart
psycopg2-binary
async-lru
ultralyticsplus==0.0.28
ultralytics
opencv-python-headless
huggingface-hub
httpx
firebase-functions
firebase-admin
a2wsgi
python-magic
